Add HitCountHelper and refactor telemetry helpers
Introduce HitCountHelper to centralize SQLite-based visitor hit counting (upsert and totals). Move TelemetricsHelper into Kgivler.Api.Helpers, improve GPU detection/error handling and adjust namespace/usings. Update Program.cs to use the new helper, clean up imports, remove inlined ProcessHitCounts, and normalize forwarded IP header handling. Minor header/comment updates in SystemCpuMonitor and file renames. These changes organize helper code and centralize telemetry/hit-count logic.
