+++ Kgivler.Api/Program.cs
@@ -1,16 +1,16 @@
 /*
- * Random Steam Game
+ * kgivler_com
  * 
  * Copyright (c) 2026 Kyle Givler
  * Licensed under the MIT License.
  */
 
-using Kgivler.Api;
 using Kgivler.Api.BackgroundServices;
+using Kgivler.Api.Helpers;
+using Microsoft.AspNetCore.HttpOverrides;
+using Microsoft.Data.Sqlite;
 using System.Diagnostics;
 using System.Runtime.InteropServices;
-using Microsoft.Data.Sqlite;
-using Microsoft.AspNetCore.HttpOverrides;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -116,11 +116,11 @@ else
 app.MapGet("/api/system/usage", async (HttpContext context) =>
 {
     // Extract the real IP address
-    var forwardedHeader = context.Request.Headers["CF-Connecting-IP"].FirstOrDefault() 
-                       ?? context.Request.Headers["X-Forwarded-For"].FirstOrDefault() 
-                       ?? context.Connection.RemoteIpAddress?.ToString() 
+    var forwardedHeader = context.Request.Headers["CF-Connecting-IP"].FirstOrDefault()
+                       ?? context.Request.Headers["X-Forwarded-For"].FirstOrDefault()
+                       ?? context.Connection.RemoteIpAddress?.ToString()
                        ?? "unknown";
-    
+
     // If multiple IPs are in X-Forwarded-For, take the first one
     var ip = forwardedHeader.Split(',')[0].Trim();
 
@@ -133,7 +133,7 @@ app.MapGet("/api/system/usage", async (HttpContext context) =>
     var cpuUsage = TelemetricsHelper.GetCpuUsage();
     var stardate = TelemetricsHelper.GetStarDate();
     var weather = await TelemetricsHelper.GetLocalWeather();
-    var hitResults = await ProcessHitCounts(connectionString, ip);
+    var hitResults = await HitCountHelper.ProcessHitCounts(connectionString, ip);
 
 
     var telemetry = new
@@ -164,41 +164,3 @@ app.MapGet("/api/system/usage", async (HttpContext context) =>
 });
 
 app.Run();
-
-
-async static Task<(long totalHits, long uniqueVisitors)> ProcessHitCounts(string connectionString, string ip)
-{
-    
-    long totalHits = 0;
-    long uniqueVisitors = 0;
-
-    // SQLite Upsert and Count
-    using var connection = new SqliteConnection(connectionString);
-    await connection.OpenAsync();
-
-    // Update the hit count
-    var upsertCmd = connection.CreateCommand();
-    upsertCmd.CommandText = @"
-            INSERT INTO Visitors (IpAddress, Hits, LastSeen)
-            VALUES ($ip, 1, $date)
-            ON CONFLICT(IpAddress) DO UPDATE SET
-                Hits = Hits + 1,
-                LastSeen = $date;
-        ";
-    upsertCmd.Parameters.AddWithValue("$ip", ip);
-    upsertCmd.Parameters.AddWithValue("$date", DateTime.UtcNow.ToString("o"));
-    await upsertCmd.ExecuteNonQueryAsync();
-
-    // Get Totals
-    var statsCmd = connection.CreateCommand();
-    statsCmd.CommandText = "SELECT COUNT(IpAddress), SUM(Hits) FROM Visitors;";
-    using var reader = await statsCmd.ExecuteReaderAsync();
-
-    if (await reader.ReadAsync())
-    {
-        uniqueVisitors = reader.IsDBNull(0) ? 0 : reader.GetInt64(0);
-        totalHits = reader.IsDBNull(1) ? 0 : reader.GetInt64(1);
-    }
-    
-    return (totalHits, uniqueVisitors);
-}
